+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### Unreleased
+
+* Replace `angular-async-upload` lib with `angular-pallet`
+
 ### v1.0.0
 
 * Initial release.